feat: update branch

+++ src/routes/album/album.controller.ts
@@ -8,33 +8,55 @@ import {
   HttpCode,
   ParseUUIDPipe,
   Put,
+  Req,
+  UnauthorizedException
 } from '@nestjs/common';
 import { AlbumService } from './album.service';
 import { ApiResponse } from '@nestjs/swagger';
 import { CreateAlbumDto } from './dto/create-album.dto';
 import { UpdateAlbumDto } from './dto/update-album.dto';
+import { Request } from 'express';
+
 
 @Controller('album')
 export class AlbumController {
   constructor(private readonly albumService: AlbumService) { }
   @ApiResponse({ status: 200, description: 'Server should answer with status code 200 and all users records' })
   @Get()
-  async findAll() {
-    return this.albumService.findAllAlbum();
+  async findAll(@Req() request: Request) {
+    // if (request.headers.authorization !== `Bearer ${process.env.BAERER_TOKEN}`) {
+    //   throw new UnauthorizedException()
+    // } else {
+      return this.albumService.findAllAlbum();
+    // }
   }
   @ApiResponse({ status: 200, description: 'Server should answer with status code 200 and and record with id === userId if it exists' })
   @ApiResponse({ status: 400, description: 'Server should answer with status code 400 and corresponding message if userId is invalid (not uuid)' })
   @ApiResponse({ status: 404, description: 'Server should answer with status code 404 and corresponding message if record with id === userId doesnt exist' })
   @Get(':id')
-  async findOne(@Param('id', new ParseUUIDPipe({ version: '4' })) id: string) {
-    return this.albumService.findOneAlbum(id);
+  async findOne(
+    @Param('id', new ParseUUIDPipe({ version: '4' })) id: string,
+    // @Req() request: Request
+  ) {
+    if (request.headers.authorization !== `Bearer ${process.env.BAERER_TOKEN}`) {
+      throw new UnauthorizedException()
+    } else {
+      return this.albumService.findOneAlbum(id);
+    }
   }
   @ApiResponse({ status: 201, description: 'Server should answer with status code 201 and newly created record if request is valid' })
   @ApiResponse({ status: 400, description: 'Server should answer with status code 400 and corresponding message if request body does not contain required fields.' })
   @Post()
   @HttpCode(201)
-  async create(@Body() createAlbumDto: CreateAlbumDto) {
-    return this.albumService.createAlbum(createAlbumDto);
+  async create(
+    @Body() createAlbumDto: CreateAlbumDto,
+    @Req() request: Request
+  ) {
+    if (request.headers.authorization !== `Bearer ${process.env.BAERER_TOKEN}`) {
+      throw new UnauthorizedException()
+    } else {
+      return this.albumService.createAlbum(createAlbumDto);
+    }
   }
   @ApiResponse({ status: 201, description: 'Server should answer with status code 200 and updated record if request is valid' })
   @ApiResponse({ status: 400, description: 'Server should answer with status code 400 and corresponding message if userId is invalid(not uuid)' })
@@ -43,17 +65,30 @@ export class AlbumController {
   async update(
     @Param('id', new ParseUUIDPipe({ version: '4' })) id: string,
     @Body() updateAlbumDto: UpdateAlbumDto,
+    @Req() request: Request
   ) {
-    await this.findOne(id);
-    return this.albumService.updateAlbum(id, updateAlbumDto);
+    if (request.headers.authorization !== `Bearer ${process.env.BAERER_TOKEN}`) {
+      throw new UnauthorizedException()
+    } else {
+      await this.findOne(id, request);
+      return this.albumService.updateAlbum(id, updateAlbumDto);
+    }
   }
   @ApiResponse({ status: 204, description: 'Server should answer with status code 200 and updated record if request is valid' })
   @ApiResponse({ status: 400, description: 'Server should answer with status code 400 and corresponding message if userId is invalid(not uuid)' })
   @ApiResponse({ status: 404, description: 'Server should answer with status code 404 and corresponding message if record with id === userId doesnt exist' })
   @Delete(':id')
   @HttpCode(204)
-  async remove(@Param('id', new ParseUUIDPipe({ version: '4' })) id: string) {
-    await this.findOne(id);
-    return this.albumService.removeAlbum(id);
+  async remove(
+    @Param('id', new ParseUUIDPipe({ version: '4' })) id: string,
+    @Req() request: Request
+
+  ) {
+    if (request.headers.authorization !== `Bearer ${process.env.BAERER_TOKEN}`) {
+      throw new UnauthorizedException()
+    } else {
+      await this.findOne(id, request);
+      return this.albumService.removeAlbum(id);
+    }
   }
 }
